+++ src/orders/dto/order.create.dto.ts
@@ -1,19 +1,7 @@
+import { Orders } from '@/orders/class/orders';
 import { ApiProperty } from '@nestjs/swagger';
 import { Type } from 'class-transformer';
-import { IsArray, IsMongoId, IsString, ValidateNested } from 'class-validator';
-
-export class Orders {
-  @ApiProperty({ type: String, description: 'menu is ObjectID' })
-  @IsMongoId()
-  menu: string;
-  @ApiProperty({ type: String, description: 'Array of ObjectID' })
-  @IsArray()
-  @IsMongoId({ each: true })
-  addons: string[];
-  @ApiProperty({ type: String })
-  @IsString()
-  additional_info: string;
-}
+import { IsArray, IsMongoId, ValidateNested } from 'class-validator';
 export class CreateOrderDto {
   @ApiProperty({ type: String, description: 'session is ObjectID' })
   @IsMongoId()
