+++ src/orders/orders.service.ts
@@ -11,12 +11,12 @@ export class OrdersService {
     @InjectModel(OrdersSchema)
     private readonly orderModel: ReturnModelType<typeof OrdersSchema>,
   ) {}
-  async createOrder(createorderdto: CreateOrderDto) {
-    const insertObject = createorderdto.orders.map((food_element) => {
+  async createOrder(createOrderDto: CreateOrderDto) {
+    const insertObject = createOrderDto.orders.map((food_element) => {
       const element = new OrdersSchema();
-      element.session = new Types.ObjectId(createorderdto.session);
+      element.session = new Types.ObjectId(createOrderDto.session);
       element.menu = new Types.ObjectId(food_element.menu);
-      element.addons = food_element.addons.map(
+      element.addons = food_element.addons?.map(
         (addons_element) => new Types.ObjectId(addons_element),
       );
       element.additional_info = food_element.additional_info;
