+++ test/orders.e2e-spec.ts
@@ -1,0 +1,25 @@
+//not finished dont't blame me.
+
+// import { OrdersModule } from '@/orders/orders.module';
+// import { OrdersService } from '@/orders/orders.service';
+// import {
+//   FastifyAdapter,
+//   NestFastifyApplication,
+// } from '@nestjs/platform-fastify';
+// import { Test } from '@nestjs/testing';
+
+// describe('orders', () => {
+//   let app: NestFastifyApplication;
+//   let orderService = { createOrder: jest.fn() };
+//   beforeAll(async () => {
+//     const moduleRef = await Test.createTestingModule({
+//       imports: [OrdersModule],
+//     })
+//       .overrideProvider(OrdersService)
+//       .useValue(orderService)
+//       .compile();
+//     app = moduleRef.createNestApplication<NestFastifyApplication>(
+//       new FastifyAdapter(),
+//     );
+//   });
+// });
